Allows manual entry of which pokeball to try and use, updates web interface to allow evolving the same way as transferring (#335)
* Allows manual entry of pokeball % to look for

This changes ```take_next_ball``` to look at the values in config.json
to decide when to switch to another ball. Because masterballs should in
theory always catch a pokemon, I disabled it by default. Setting
```use_masterball_at``` to anything other than 0 will result in an
attempt to use a masterball if the ultraball % isn’t met. If this is
set to 0 and ultraball % still isn’t met, ```get_best_ball``` will
return an ultraball.

* Updated the terminology to be more clear

Also rewrote it so that it’s uppercase.

* Shouldn’t we use a lucky egg if applicable before attempting to evolve
for more xp?